- Moved porevious sim to legacy folder - Added new sim that utilizes SimPy
